requirements.txt
python-dotenv
python-telegram-bot==12.8